feat: upload the data of the number of users
