+++ src/main.js
@@ -21,7 +21,7 @@ import directive from '@/directive'
 
 import App from './App.vue'
 import { analyticsRequestAPI } from '@api'
-import { JSEncrypt } from 'jsencrypt'
+import Encrypt from './utilities/encrypt'
 import { NativeEventSource, EventSourcePolyfill } from 'event-source-polyfill'
 global.EventSource = EventSourcePolyfill || NativeEventSource
 
@@ -59,30 +59,14 @@ const userName = () => {
 const analyticsRequest = (to, from) => {
   const hostname = window.location.hostname
   if (to.path !== from.path && !utils.isPrivateIP(hostname)) {
-    const publicKey = `-----BEGIN RSA PUBLIC KEY-----
-    MIIBpTANBgkqhkiG9w0BAQEFAAOCAZIAMIIBjQKCAYQAz5IqagSbovHGXmUf7wTB
-    XrR+DZ0u3p5jsgJW08ISJl83t0rCCGMEtcsRXJU8bE2dIIfndNwvmBiqh13/WnJd
-    +jgyIm6i1ZfNmf/R8pEqVXpOAOuyoD3VLT9tfWvz9nPQbjVI+PsUHH7nVR0Jwxem
-    NsH/7MC2O15t+2DVC1533UlhjT/pKFDdTri0mgDrLZHp6gPF5d7/yQ7cPbzv6/0p
-    0UgIdStT7IhkDfsJDRmLAz09znv5tQQtHfJIMdAKxwHw9mExcL2gE40sOezrgj7m
-    srOnJd65N8anoMGxQqNv+ycAHB9aI1Yrtgue2KKzpI/Fneghd/ZavGVFWKDYoFP3
-    531Ga/CiCwtKfM0vQezfLZKAo3qpb0Edy2BcDHhLwidmaFwh8ZlXuaHbNaF/FiVR
-    h7uu0/9B/gn81o2f+c8GSplWB5bALhQH8tJZnvmWZGI9OnrIlWmQZsuUBooTul9Q
-    ZJ/w3sE1Zoxa+Or1/eWijqtIfhukOJBNyGaj+esFg6uEeBgHAgMBAAE=
-    -----END RSA PUBLIC KEY-----`
     const data = {
       domain: hostname,
       username: userName(),
       url: to.path,
       createdAt: Math.floor(Date.now() / 1000)
     }
-    const encryptor = new JSEncrypt()
-    encryptor.setPublicKey(publicKey)
-    const encryptData = encryptor.encrypt(JSON.stringify(data))
-    const payload = {
-      data: encryptData
-    }
-    analyticsRequestAPI(payload)
+
+    analyticsRequestAPI(Encrypt(data))
       .then()
       .catch((err) => {
         console.log(err)
